requirements.txt
matplotlib==3.3.0
pandas==1.1.0
torchvision==0.1.8
pytesseract==0.3.4
Pillow==7.2.0